requirements/requirements.txt
fastapi[standard]==0.128.8
geoip2==5.1.0
uvicorn==0.27.1
python-dotenv==1.2.1
